Added some flow control support
The serial port on the Arduino doesn't have flow control.  The pins
aren't even wired up.  I've added some flow control to the serial line.
What that means is that it sends an xoff when it's going to do something
and xon when it's back from that.  Maybe I should try xing-off when the
pipeline fills and then xing-on when it's less full.  Anyway, the
high-level flow control is working, it's the low-level stuff that's
busted.