Add methods to save and fetch the configuration.  Also do the transmitter select buttons.
Even though I use red and green to indicate which transmitter is selected, it should be different
enough for the color blind.  I also use the relief to indicate which button the mouse cursor is over.
Although the effect is subtle, it should be sufficient.

With respect to the plumbing for the configuration load/restore, the plan is to eventually have each
class return a Python structure that it knows how to create from the actual instance's configuration
and build up a hierarchy of such structures.  That would be converted into JSON at the top level,
which could then be saved to a file.  When reading the saved configuration, the process would be
reversed.