Change what the text-to-morse method returns and work on the memories
Previously, I had the MorseEncoder.one_letter method sending messages directly
to the dongle.  That is, the content of the message was specified in that method.
This is obviously wrong.  The MorseEncoder should know nothing about how the
program as a whole talks to the dongle, it should only return some sort of Morse
representation of the character that is passed in.  This, of course, required that
I change the keystroke handling to deal with the changes, so I did.  Last, I added
code to the memory processor to send the content of the memory when the memory
button is clicked.  Currently, they all contain "R" but that will change.